computer-vision/segment_anything/webapp/requirements.txt
requests==2.31.0
opencv-python
numpy
Pillow
streamlit
streamlit-image-coordinates
streamlit-dimensions
